config/runtime: enable 'local_src' debug option in python.jinja2
Allow building kernels that are already extracted to a local dir instead
of downloading a new tarball every time.
This can be enabled for development by including a 'local_src' key in
the 'debug' node attribute containing the local path of the kernel
source.

The default behavior if this debug option is not enabled is still the
same.

Signed-off-by: Ricardo Cañuelo <ricardo.canuelo@collabora.com>